--- a/src/engine/source/base/utils/stringUtils.hpp
+++ b/src/engine/source/base/utils/stringUtils.hpp
@@ -10,20 +10,71 @@
 #ifndef _STRING_UTILS_H
 #define _STRING_UTILS_H
 
-#include <vector>
 #include <string>
+#include <vector>
 
-namespace utils::string {
+namespace utils::string
+{
 
-    /**
-     * @brief Split a string into a vector of strings
-     *
-     * @param str String to be split
-     * @param delimiter Delimiter to split the string
-     * @return std::vector<std::string>
-     */
-    std::vector<std::string> split(std::string_view str, char delimiter);
+/**
+ * @brief Split a string into a vector of strings
+ *
+ * @param str String to be split
+ * @param delimiter Delimiter to split the string
+ * @return std::vector<std::string>
+ */
+std::vector<std::string> split(std::string_view str, char delimiter);
 
+using Delimeter = std::pair<char, bool>;
+
+/**
+ * @brief
+ *
+ * @tparam Delimiter Pair with delimiter char and a boolean indicating if the
+ * delimiter should also be included in the result
+ * @param input Input string
+ * @param delimiters Delimiters to split the string
+ * @return std::vector<std::string> Vector of strings
+ */
+template<typename... Delimiter>
+std::vector<std::string> split(const std::string &input,
+                               Delimiter &&...delimiters)
+{
+    std::vector<std::string> splitted;
+    for (auto i = 0, last = i; i < input.size(); ++i)
+    {
+        for (auto delimiter : {delimiters...})
+        {
+            if (input[i] == delimiter.first)
+            {
+                auto substr = input.substr(last, i - last);
+                if (!substr.empty())
+                {
+                    splitted.push_back(substr);
+                }
+
+                if (delimiter.second)
+                {
+                    splitted.push_back({delimiter.first});
+                }
+
+                last = i + 1;
+                break;
+            }
+        }
+        if (i == input.size() - 1)
+        {
+            auto substr = input.substr(last);
+            if (!substr.empty())
+            {
+                splitted.push_back(substr);
+            }
+        }
+    }
+
+    return splitted;
 }
+
+} // namespace utils::string
 
 #endif // _STRING_UTILS_H