--- a/src/engine/source/builder/builders/opBuilderCondition.hpp
+++ b/src/engine/source/builder/builders/opBuilderCondition.hpp
@@ -16,6 +16,37 @@ namespace builder::internals::builders
 {
 
 /**
+ * @brief Builds condition value operation.
+ * Checks that a tuple <field: value> is present in the event.
+ *
+ * @param def Definition of the operation to be built
+ * @return types::Lifter
+ */
+std::function<bool(base::Event)>
+middleBuilderConditionReference(const base::DocumentValue& def,
+                                types::TracerFn tr);
+
+/**
+ * @brief Builds condition value operation.
+ * Checks that a tuple <field: value> is present in the event.
+ *
+ * @param def Definition of the operation to be built
+ * @return types::Lifter
+ */
+std::function<bool(base::Event)>
+middleBuilderConditionValue(const base::DocumentValue& def,
+                            types::TracerFn tr);
+
+/**
+ * @brief Auxiliray builder to handle proper specific condition builder
+ *
+ * @param def Definition of the operation to be built
+ * @return base::Lifter
+ */
+std::function<bool(base::Event)>
+middleBuilderCondition(const base::DocumentValue& def, types::TracerFn tr);
+
+/**
  * @brief Auxiliray builder to handle proper specific condition builder
  *
  * @param def Definition of the operation to be built