--- a/src/engine/source/builder/definitions.hpp
+++ b/src/engine/source/builder/definitions.hpp
@@ -1,0 +1,16 @@
+#ifndef _DEFINITIONS_H_
+#define _DEFINITIONS_H_
+
+#include "builderTypes.hpp"
+
+namespace builder
+{
+namespace internals
+{
+
+void substituteDefinitions(base::Document& asset);
+
+}
+} // namespace builder
+
+#endif // _DEFINITIONS_H_