… src/engine/CMakeLists.txt
… src/engine/benchmark/CMakeLists.txt
… src/engine/benchmark/logicExpression/CMakeLists.txt
… src/engine/benchmark/logicExpression/logicExpressionEvaluator_bench.cpp
… src/engine/source/builder/CMakeLists.txt
… src/engine/source/builder/builders/opBuilderConditionReference.cpp
… src/engine/source/builder/builders/opBuilderConditionReference.hpp
… src/engine/source/builder/builders/opBuilderConditionValue.cpp
… src/engine/source/builder/builders/opBuilderConditionValue.hpp
… src/engine/source/builder/builders/opBuilderHelperFilter.cpp
… src/engine/source/builder/builders/opBuilderHelperFilter.hpp
… src/engine/source/builder/register.hpp
… src/engine/source/logicExpression/CMakeLists.txt
… src/engine/source/logicExpression/include/logicExpression/logicExpression.hpp
… src/engine/source/logicExpression/src/logicExpressionEvaluator.hpp
… src/engine/source/logicExpression/src/logicExpressionParser.cpp
… src/engine/source/logicExpression/src/logicExpressionParser.hpp
… src/engine/test/CMakeLists.txt
… src/engine/test/source/base/stringUtils_test.cpp
… src/engine/test/source/builder/CMakeLists.txt
… src/engine/test/source/builder/builders/assetBuilderDecoder_test.cpp
… src/engine/test/source/builder/builders/assetBuilderFilter_test.cpp
… src/engine/test/source/builder/builders/assetBuilderOutput_test.cpp
… src/engine/test/source/builder/builders/assetBuilderRule_test.cpp
… src/engine/test/source/builder/builders/opBuilderConditionReference_test.cpp
… src/engine/test/source/builder/builders/opBuilderConditionValue_test.cpp
… src/engine/test/source/builder/builders/opBuilderCondition_test.cpp
… src/engine/test/source/builder/builders/opBuilderHelperExists_test.cpp
… src/engine/test/source/builder/builders/opBuilderHelperIPCIDRCheck_test.cpp
… src/engine/test/source/builder/builders/opBuilderHelperIntEqual_test.cpp