|Related issue|
|---|
|Closes #13008 |

We want to give the user the ability to express complex logical expressions to decide whether an event can be processed by an asset, or as a condition to execute a mapping.

This PR adds the conditional expression stage to the Engine. Adds a new module to generate a function to evaluate a logical expression composed of check operations. Adds the builder for the new stage and refactors current conditional builders to return a function instead of the rxcpp observable. 

# Tests (local)
![image](https://user-images.githubusercontent.com/32262972/169838530-9518f08d-8b70-4e17-b06f-5ca38a765b91.png)

# Benchmark
### PR - 3 threads (1 processing - 1 server - 1 logger)(Tracing deactivated)(Release)
main command: `./main -f ../test/assets/ -e tcp:localhost:5054 -t 1 --kvdbPath /tmp`
benchmark command: `go run benchmark_tool.go -o /tmp/filepath.txt -p tcp -s 127.0.0.1:5054 -t 60  -r 0 -b`
```
Benchmark report
----------------

Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.103324 seconds
Sent events:         2255155
Processed events:    2255155
Lost events:               0
EPS:                37585.91
```
pidstat snapshot under heavy load `pidstat 1 --human -G main -t -u -r -h`:
```
# Time        UID      TGID       TID    %usr %system  %guest   %wait    %CPU   CPU  minflt/s  majflt/s     VSZ     RSS   %MEM  Command
14:31:03     1000     19146         -  193.0%    9.0%    0.0%    1.0%  202.0%     5      4.00      0.00  145.8M   49.3M   0.3%  main
14:31:03     1000         -     19146  100.0%    0.0%    0.0%    1.0%  100.0%     5      4.00      0.00  145.8M   49.3M   0.3%  |__main
14:31:03     1000         -     19147    0.0%    2.0%    0.0%    2.0%    2.0%     1      0.00      0.00  145.8M   49.3M   0.3%  |__main
14:31:03     1000         -     19148   91.0%    9.0%    0.0%    0.0%  100.0%     6      0.00      0.00  145.8M   49.3M   0.3%  |__main
```
### Dev - 3 threads (1 processing - 1 server - 1 logger)(Tracing deactivated)(Release)
main command: `./main -f ../test/assets/ -e tcp:localhost:5054 -t 1 --kvdbPath /tmp`
benchmark command: `go run benchmark_tool.go -o /tmp/filepath.txt -p tcp -s 127.0.0.1:5054 -t 60  -r 0 -b`
```
Benchmark report
----------------

Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.156680 seconds
Sent events:         2238040
Processed events:    2238040
Lost events:               0
EPS:                37300.66
```
pidstat snapshot under heavy load `pidstat 1 --human -G main -t -u -r -h`:
```
# Time        UID      TGID       TID    %usr %system  %guest   %wait    %CPU   CPU  minflt/s  majflt/s     VSZ     RSS   %MEM  Command
14:38:40     1000     21563         -  189.0%   13.0%    0.0%    0.0%  202.0%     4      0.00      0.00  145.9M   49.7M   0.3%  main
14:38:40     1000         -     21563  100.0%    0.0%    0.0%    0.0%  100.0%     4      0.00      0.00  145.9M   49.7M   0.3%  |__main
14:38:40     1000         -     21564    0.0%    2.0%    0.0%    1.0%    2.0%     3      0.00      0.00  145.9M   49.7M   0.3%  |__main
14:38:40     1000         -     21565   89.0%   11.0%    0.0%    0.0%  100.0%     2      0.00      0.00  145.9M   49.7M   0.3%  |__main
```